refactor and style profile
